Linux/ROCm: build the tvm-ffi kernels on gfx1100
The ROCm port was Windows-only in two places that blocked every engine
kernel on Linux ROCm. Verified on an RX 7900 XTX (gfx1100, ROCm 7.2.3,
torch 2.10.0+rocm7.0): tests/moe/test_fused_copy.py and
tests/kernels/test_pinned_tensor.py go from 6 failures to 1 (the one
left needs the _pinned_tensor extension, which setup.py never builds).

* kernel/utils.py: pick the device-compiler flags from torch.version.hip
  rather than HIP_PATH. HIP_PATH is the Windows/TheRock convention and is
  normally unset on Linux ROCm, so the nvcc-only --expt-relaxed-constexpr
  was handed to hipcc and every tvm-ffi kernel died with
  "clang++: error: unknown argument". The env vars stay as a fallback for
  the AOT cache build, which runs without importing torch.

* csrc: a ROCm build reports tensors as rocm / rocm_host, so matchers
  hard-coded to kDLCUDA rejected every tensor with "Device value
  [rocm[0]] not in the allowed options: [cuda, cuda_host, cpu]". Add
  host::kFTDeviceGPU / kFTDeviceGPUHost, resolved per-platform from
  __HIP_PLATFORM_AMD__ / USE_ROCM, and use them in fast_index_copy's
  device_alias + matchers and in pynccl's device checks.

* .gitignore: torch's cpp_extension writes hipify output next to the
  .cu/.cuh sources; those are build artifacts.

+++ python/freetoken/kernel/csrc/gguf/dispatch.h
@@ -11,12 +11,16 @@
 #endif
 
 // Warp-shuffle wrappers the donor pulls from sgl-kernel's utils.h (CUDA variants).
+// All call sites pass a 32-bit mask literal (uint32_t(-1)); CUDA's __shfl_xor_sync
+// accepts that, but ROCm's __shfl_xor_sync static_asserts the mask is a genuine
+// 64-bit type (AMD wavefronts can be 64 lanes) -- widen it here so every caller is
+// covered without touching each one.
 #ifndef SGLANG_SHFL_XOR_SYNC
-#define SGLANG_SHFL_XOR_SYNC(mask, var, lane_mask) __shfl_xor_sync((mask), (var), (lane_mask))
+#define SGLANG_SHFL_XOR_SYNC(mask, var, lane_mask) __shfl_xor_sync((uint64_t)(mask), (var), (lane_mask))
 #endif
 #ifndef SGLANG_SHFL_XOR_SYNC_WIDTH
 #define SGLANG_SHFL_XOR_SYNC_WIDTH(mask, var, lane_mask, width) \
-  __shfl_xor_sync((mask), (var), (lane_mask), (width))
+  __shfl_xor_sync((uint64_t)(mask), (var), (lane_mask), (width))
 #endif
 
 #define DISPATCH_CASE_FLOAT_TYPES(...)                 \
